Extract logic into custom hook
